+++ Localizable.xcstrings
@@ -490,10 +490,7 @@
     "チームカラー選択" : {
 
     },
-    "チームプリセット作成" : {
-
-    },
-    "チームプリセット確認" : {
+    "チームプリセット一覧" : {
 
     },
     "チーム名" : {
