+++ RugStats/View/AdvanceModeSetting/AdvanceModeSettingView.swift
@@ -13,13 +13,11 @@ struct AdvanceModeSettingView: View {
     var body: some View {
         NavigationStack{
             VStack{
-                NavigationLink("チームプリセット作成", destination: CreateTeamPresetView())
-                
                 // TODO: 遷移先画面を作成
                 NavigationLink("アクションプリセット作成", destination: EmptyView())
                 
                 // TODO: プリセット一覧画面を作成
-                NavigationLink("チームプリセット確認", destination: TeamPresetListView(teamList: $teamList))
+                NavigationLink("チームプリセット一覧", destination: TeamPresetListView(teamList: $teamList))
             }
         }
     }
